Display unsupported message for IE 10 and below
Detect versions of IE 10 and below and serve up an unsupported browser
message instead of the normal page. Unfortunately, this message doesn't
work in ie 8 and below because of how the content of the page is loaded.
Instead, the page doesn't load at all.

Limited styling has also been applied to the message.